Profile with document validation

+++ app/Http/Controllers/Front/ProfileManagementController.php
@@ -10,49 +10,70 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
+
 
 
 class ProfileManagementController extends Controller
 {
-    public function updateProfile(Request $request){
+    public function updateProfile(Request $request)
+    {
         // dd($request->all());
         $data = [];
         $msg = '';
-        if($request->personal_informations =="personal_informations" && $request->profile_type == "instructor_profile"){
-            
-            array_push($data,[
-                'name'=>$request->fname,
-                'last_name'=>$request->lname,
-                'email'=> $request->email,
+        if ($request->personal_informations == "personal_informations" && $request->profile_type == "instructor_profile") {
+            $numberExists = User::where('phone', $request->phone)
+                ->where('id', '!=', $request->id) // Exclude the current user
+                ->exists();
+
+            if ($numberExists) {
+                // The number is already used by other users
+                return redirect()->back()->with('error' , 'This phone number is already taken by another user.');
+            }
+            $rules = [
+                'fname' => 'required',
+                'email' => ['required', 'email'],
+                'phone' => 'required',
+                'phone_code' => 'required',
+            ];
+
+            $customMessages = [
+                'required' => 'The :attribute field is required.'
+            ];
+            $this->validate($request, $rules, $customMessages);
+            array_push($data, [
+                'name' => $request->fname,
+                'last_name' => $request->lname,
+                'email' => $request->email,
                 'phone' => $request->phone,
-                'country_code' =>$request->phone_code,
-                'country'=>$request->country
+                'country_code' => $request->phone_code,
+                'country' => $request->country
             ]);
             $msg = 'Your profile personal information updated successfully.';
         }
 
-        if($request->personal_informations =="instructor_profile" && $request->profile_type == "instructor_profile"){
+        if ($request->personal_informations == "instructor_profile" && $request->profile_type == "instructor_profile") {
             $documentTypes = $request->document_name;
-            if($request->facebook != null || $request->instagram || $request->youtube || $request->linkedin || $request->twitter){
-                
-                 $rules = [
-                    'facebook' => ['required', new ValidUrl],
-                    'instagram' => ['required', new ValidUrl],
-                    'youtube' => ['required', new ValidUrl],
-                    'linkedin' => ['required', new ValidUrl],
-                    'twitter' => ['required', new ValidUrl],
+            if ($request->facebook != null || $request->instagram || $request->youtube || $request->linkedin || $request->twitter) {
+
+                $rules = [
+                    'facebook' => ['nullable', new ValidUrl],
+                    'instagram' => ['nullable', new ValidUrl],
+                    'youtube' => ['nullable', new ValidUrl],
+                    'linkedin' => ['nullable', new ValidUrl],
+                    'twitter' => ['nullable', new ValidUrl],
                 ];
-            
+
                 $customMessages = [
                     'required' => 'The :attribute field is required.'
                 ];
-            
+
                 $this->validate($request, $rules, $customMessages);
             }
-            array_push($data,[
-                'qualification'=>$request->qualifications,
-                'workplace'=>$request->profession_work,
-                'teaching_experience'=> $request->teaching_mentorship,
+            array_push($data, [
+                'qualification' => $request->qualifications,
+                'workplace' => $request->profession_work,
+                'teaching_experience' => $request->teaching_mentorship,
                 'summary' => $request->about_me,
                 'Facebook' => $request->facebook,
                 'Instagram' => $request->instagram,
@@ -60,82 +81,94 @@ class ProfileManagementController extends Controller
                 'LinkedIn' => $request->linkedin,
                 'Twitter' => $request->twitter,
             ]);
-            if(count($documentTypes) > 0){
-                foreach($documentTypes as $key => $documentType){
-                    
+            if (count($documentTypes) > 0) {
+                $errors = [];
+                foreach ($documentTypes as $key => $documentType) {
+
                     $document_id = $request->document_id[$key] ?? null;
 
                     $old_document = $request->old_document[$key];
 
-                    $documentArray =[
-                        'document_name' =>$documentType
+                    if($documentType != null && !isset($request->supporting_document[$key]) && $old_document == null && $document_id == null){
+                        // dd($old_document,$request->all());
+                        return redirect()->back()->with('error' , 'Please add document.');
+                    }
+
+                    // Skip validation if old_document is present and supporting_document is empty
+                    
+                    $documentArray = [
+                        'document_name' => $documentType
                     ];
-                    if($document_id != null){
+                    if ($document_id != null) {
                         if ($old_document != null && $document_id != null) {
                             // print_r($request->supporting_document[$key]);
                             if (isset($request->supporting_document[$key])) {
                                 Storage::delete('public/documents/' . $old_document);
                                 $documentFile = $request->supporting_document[$key];
-    
-                                $str=rand();
+
+                                $str = rand();
                                 $string = md5($str);
-                                $documentFileName = $string.'.'.$documentFile->getClientOriginalExtension();
+                                $documentFileName = $string . '.' . $documentFile->getClientOriginalExtension();
                                 $documentFile->storeAs('public/documents', $documentFileName);
                                 $documentArray['user_document'] = $documentFileName;
                             }
                         }
-                        User_documents::where('id','=',$document_id)->update($documentArray);
-                    }else{
-                        $supporting_document = $request->supporting_document[$key];
-                        $documentSave = new User_documents();
-                        $documentSave->document_name = $documentType;
-                        $documentSave->user_id = $request->id;
-                        if ($supporting_document) {
-                            $documentFile = $supporting_document;
+                        User_documents::where('id', '=', $document_id)->update($documentArray);
+                    } else {
+                        if (isset($request->supporting_document[$key])) {
+                            $supporting_document = $request->supporting_document[$key];
+                            $documentSave = new User_documents();
+                            $documentSave->document_name = $documentType;
+                            $documentSave->user_id = $request->id;
+                            if ($supporting_document) {
+                                $documentFile = $supporting_document;
 
-                            $str=rand();
-                            $string = md5($str);
-                            $documentFileName = $string.'.'.$documentFile->getClientOriginalExtension();
-                            $documentFile->storeAs('public/documents', $documentFileName);
-                            $documentSave->user_document = $documentFileName;
+                                $str = rand();
+                                $string = md5($str);
+                                $documentFileName = $string . '.' . $documentFile->getClientOriginalExtension();
+                                $documentFile->storeAs('public/documents', $documentFileName);
+                                $documentSave->user_document = $documentFileName;
+                            }
+                            $documentSave->save();
                         }
-                        $documentSave->save();
-                        
                     }
                 }
                 $msg = 'Your instructor profile updated successfully.';
             }
         }
         // exit;
-        User::where('id','=',$request->id)->update($data[0]);
+        User::where('id', '=', $request->id)->update($data[0]);
         return redirect()->back()->with('success', $msg);
 
     }
 
-    public function updateProfileImage(Request $request){
+    public function updateProfileImage(Request $request)
+    {
         $request->validate([
             'avtar' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048', // Customize the rules based on your requirements
         ]);
         if ($request->file('avtar')) {
             Storage::delete('public/users/' . $request->old_avtar);
             $image = $request->file('avtar');
-            $imageName = time().'.'.$image->getClientOriginalExtension();
+            $imageName = time() . '.' . $image->getClientOriginalExtension();
 
             // Store the image in the storage directory (you can specify a subdirectory if needed)
             $image->storeAs('public/users', $imageName);
-            User::where('id','=',$request->id)->update(['avatar'=>$imageName]);
+            User::where('id', '=', $request->id)->update(['avatar' => $imageName]);
         }
         return redirect()->back()->with('success', 'Avatar uploaded successfully!');
     }
 
-    public function deleteDocument(Request $request){
-        $deleteDocument = User_documents::where('id','=',$request->id)->first();
+    public function deleteDocument(Request $request)
+    {
+        $deleteDocument = User_documents::where('id', '=', $request->id)->first();
         Storage::delete('public/documents/' . $deleteDocument->user_document);
         $deleteDocument->delete();
-        return response()->json(["message"=>"Document deleted successfully...!","status"=>200],200);
+        return response()->json(["message" => "Document deleted successfully...!", "status" => 200], 200);
     }
 
-    public function changePassword(Request $request){
+    public function changePassword(Request $request)
+    {
         // dd(Hash::make('12345678'));
         $request->validate([
             'old_password' => 'required',
